Add a test suite for in-memory mode
